requirements.txt
streamlit
torch
torchvision
ultralytics
sahi
pillow
numpy
matplotlib
